+++ api-scanner/src/main/java/net/corda/plugins/ApiScanner.java
@@ -4,14 +4,19 @@ import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.ConfigurationContainer;
 import org.gradle.api.file.FileCollection;
+import org.gradle.api.provider.Provider;
 import org.gradle.api.tasks.TaskCollection;
 import org.gradle.api.tasks.TaskProvider;
 import org.gradle.jvm.tasks.Jar;
 
 import javax.annotation.Nonnull;
+import java.util.Objects;
 
 @SuppressWarnings("unused")
 public class ApiScanner implements Plugin<Project> {
+    private static final String CLASSIFIER_PROPERTY_NAME = "cordaScanApiClassifier";
+    private static final String DEFAULT_CLASSIFIER = "";
+    private static final String SCAN_TASK_NAME = "scanApi";
 
     /**
      * Identify the Gradle Jar tasks creating jars
@@ -22,15 +27,21 @@ public class ApiScanner implements Plugin<Project> {
     @Override
     public void apply(@Nonnull Project project) {
         project.getLogger().info("Applying API scanner to {}", project.getName());
+        String targetClassifier = (String) project.findProperty(CLASSIFIER_PROPERTY_NAME);
+        if (targetClassifier == null) {
+            targetClassifier = DEFAULT_CLASSIFIER;
+        }
 
-        ScannerExtension extension = project.getExtensions().create("scanApi", ScannerExtension.class);
+        ScannerExtension extension = project.getExtensions().create(SCAN_TASK_NAME, ScannerExtension.class, targetClassifier);
 
         // Register the scanning task lazily, so that it will be configured after the project has been evaluated.
-        project.getLogger().info("Adding scanApi task to {}", project.getName());
-        TaskProvider<ScanApi> scanProvider = project.getTasks().register("scanApi", ScanApi.class, scanTask -> {
+        project.getLogger().info("Adding {} task to {}", SCAN_TASK_NAME, project.getName());
+        TaskProvider<ScanApi> scanProvider = project.getTasks().register(SCAN_TASK_NAME, ScanApi.class, scanTask -> {
             TaskCollection<Jar> jarTasks = project.getTasks()
                 .withType(Jar.class)
-                .matching(jarTask -> jarTask.getClassifier().isEmpty() && jarTask.isEnabled());
+                .matching(jarTask ->
+                     matches(jarTask.getArchiveClassifier(), extension.getTargetClassifier()) && jarTask.isEnabled()
+                );
             FileCollection jarSources = project.files(jarTasks);
 
             scanTask.setClasspath(compilationClasspath(project.getConfigurations()));
@@ -48,6 +59,10 @@ public class ApiScanner implements Plugin<Project> {
             target.getTasks().withType(GenerateApi.class, generateTask -> generateTask.dependsOn(scanProvider));
             target = target.getParent();
         }
+    }
+
+    private static boolean matches(Provider<String> a, Provider<String> b) {
+        return Objects.equals(a.getOrNull(), b.getOrNull());
     }
 
     private static FileCollection compilationClasspath(ConfigurationContainer configurations) {